+++ brest/supplies/__init__.py
@@ -12,7 +12,8 @@
 from .supplies import Supplies
 from .mansup import Mansup
 from .tenma import Tenma
+from .mp71 import MP71
 
-__all__ = ['Supplies', 'Mansup', 'Tenma']
+__all__ = ['Supplies', 'Mansup', 'Tenma', 'MP71']
 
 __version__ = '1.0.0'